requirements.txt
python-telegram-bot==20.8
aiohttp==3.9.1
python-dotenv==1.0.0
flask==3.0.0
